tests/wpt/tests/webmcp/imperative/non-secure.html

<!DOCTYPE html>
<html>
<head>
<title>WebMCP restricted to secure contexts</title>
<link rel="author" href="mailto:dom@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(() => {
  assert_false(document.hasOwnProperty('modelContext'));
}, `document.modelContext is not supported in non-secure context.`);
</script>
</body>
</html>
